fix description and advice handling of newlines
Now the import of desc and advice is different and both show an extra
newline in the GUI. This fixes this behavior - GUI and control file will
show the same.
The advice did not have a strip on the later write. This causes an extra
newline per load and save workflow. This ist fixed also with this changes.