fix automount on systems with disconnected (persistent) net drives
If the current user has persistent drives, but they could not be connected
on startup (e.g. no network to vpn) they shown as free by kernel32.GetLogicalDrives().
This patch will check for persistent drives and add them to the given
drive_bitmask. Than it's possible to automount to e.g. w if x,y and z are
disconnected drives (shown by the explorer with a red X).
There are two ways of the implementation. By registry (small and easy) and
per win32 API. The later way is recommanded on many site, but I could not
find any different or advantage, but leave both implementations.
The default would be the registry.